Before ReStructuring...

+++ OpenGlTraining/src/tests/TestTexture2D.cpp
@@ -52,8 +52,8 @@ void test::TestTexture2D::OnRender()
 	//m_Rotation = (float)glfwGetTime() * 25.0f;
 	m_Rotation = 1.0f;
 	const float radius = 10.0f;
-	float camX = sin(glfwGetTime()) * radius;
-	float camZ = cos(glfwGetTime()) * radius;
+	float camX = sin((float)glfwGetTime()) * radius;
+	float camZ = cos((float)glfwGetTime()) * radius;
 	glm::mat4 view;
 	view = glm::lookAt(glm::vec3(camX, 0.0, camZ), glm::vec3(0.0, 0.0, 0.0),
 		glm::vec3(0.0, 1.0, 0.0));
